settings_loxone: Filter Miniservers and Controls not in LB
Filter Miniserver list, and Controls to only match Miniservers defined in LoxBerry
If an Loxone XML includes a Miniserver (e.g. Gateway/Concentrator) that is not configured in LoxBerry, that Miniserver and controls now are not displayed anymore.

This is a UI-only filter. The controls of "foreign" Miniservers still are present in the temporary LoxPlan json files

Closes #75